Bug Fix: Create a get_player_id function to get player id
Since different players can have the same web name, the
PLAYER_WEB_NAME_TO_DICT hash wouldn't have unique keys for each player.
So, the workaround for now is to map player web name to a list of
players with that webname and getting the proper player id based on
their team name.